+++ server/src/drivers/mod.rs
@@ -1,5 +1,6 @@
 pub mod files;
 pub mod kafka;
+pub mod otlp;
 pub mod redis;
 pub mod tcp;
 pub mod unix;